3 lab

+++ lab1/src/main/java/danyliuk/mykola/model/Polynomial.java
@@ -10,6 +10,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class Polynomial {
 
-    private double[] coefficients;
+    private int[] coefficients;
 
 }
